% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_its_model.R
\name{fit_its_model}
\alias{fit_its_model}
\title{fit_its_model}
\usage{
fit_its_model(
  transformed_data,
  impact_model,
  num_interventions,
  method = "REML",
  grid_search = TRUE,
  p = NULL,
  q = NULL,
  return_grid_search = FALSE,
  ...
)
}
\arguments{
\item{transformed_data}{Am unmodified data frame created from `transform_data()`.}

\item{impact_model}{The hypothesized impact model from interventions. Available options are 'level', 'slope', or 'levelslope'.}

\item{num_interventions}{The number of interventions in your transformed data. Should be the vector length of `intervention_dates` passed in `transform_data()`.}

\item{method}{The estimation method for `nlme::gls()`, either "REML" (default) or "ML".}

\item{grid_search}{logical for whether to perform a grid search for determining lag parameters (p = AR, q = MA). By default, a grid up to values of 5 for each parameter is searched.}

\item{p}{The order of the autoregressive component. Defaults to `NULL`. If `grid_search is enabled`, this argument is ignored.}

\item{q}{The order of the moving average component. Defaults to `NULL`. If `grid_search is enabled`, this argument is ignored.}

\item{return_grid_search}{Logical flag returns the result of the grid search instead of the model. 'FALSE' by default.}

\item{...}{Additional arguments passed to `nlme::gls()`.}
}
\value{
A `gls` object of the fitted model.
}
\description{
Fits an interrupted time series model using the \CRANpkg{nlme} package, defaulting to an autocorrelation-moving average correlation structure of order (p, q)
}
