% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_predictions.R
\name{generate_predictions}
\alias{generate_predictions}
\title{Generate Predictions from a Transformed Dataset and a Model}
\usage{
generate_predictions(transformed_data, model)
}
\arguments{
\item{transformed_data}{A data frame containing the transformed dataset. Output from `transform_data()`.}

\item{model}{A GLS model object used to make predictions. The model should include coefficients used to generate pre-intervention predictions.}
}
\value{
A data frame with the following additional columns:
\describe{
  \item{pre_intervention_predictions}{Predicted values for the pre-intervention period, calculated using model coefficients.}
  \item{predictions}{Predicted values for the entire dataset using the GLS model.}
  \item{se}{Standard errors for the predictions.}
}
}
\description{
This function generates predictions and standard errors for a given transformed dataset using the provided generalized least squares (GLS) model. It also calculates pre-intervention predictions based on model coefficients.
}
\details{
The function first computes pre-intervention predictions using model coefficients, specifically the intercept, `time_index`, and interaction terms. It ensures that predictions for the pre-intervention period (`level_pre_intervention == 1`) are set to `NA`. Then, it uses the `AICcmodavg::predictSE.gls()` function to calculate predictions and standard errors from the provided GLS model.
}
\examples{
\dontrun{
  # Assuming `transformed_data` is a prepared data frame and `model` is a GLS model:
  predictions_df <- generate_predictions(transformed_data, model)
  head(predictions_df)
}

}
