% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{its_data_gp}
\alias{its_data_gp}
\title{GP programme Data}
\format{
## `its_data_gp`
A data frame with 104 rows and 4 columns:
\describe{
  \item{Date}{Weekly dates}
  \item{group_var}{Variable to indicate either treatment or control row of interest}
  \item{Period}{Description of intervention phase}
  \item{score}{The outcome metric of the scenario - self-reported wellbeing score"}
  ...
}
}
\usage{
its_data_gp
}
\description{
A dummy data set documenting a two stage level intervention programme of the medical GP practice scenario in the vignette.
Report ...
}
\keyword{datasets}
