% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{its_data_school}
\alias{its_data_school}
\title{Dummy School Data}
\format{
## `its_data_school`
A data frame with 104 rows and 4 columns:
\describe{
  \item{Date}{Weekly dates}
  \item{group_var}{Variable to indicate either treatment or control row of interest}
  \item{Period}{Description of intervention phase}
  \item{score}{The outcome metric of the senario - reading comprehension score"}
  ...
}

## `its_data_bike_programme`
A data frame with 104 rows and 4 columns:
\describe{
  \item{Date}{Weekly dates}
  \item{group_var}{Variable to indicate either treatment or control row of interest}
  \item{Period}{Description of intervention phase}
  \item{score}{The outcome metric of the scenario - weekly total maintenance incidents"}
  ...
}
}
\usage{
its_data_school

its_data_school
}
\description{
A dummy data set documenting a two stage intervention programme of the school scenario in the vignette.
Report ...

A dummy data set documenting a two stage level intervention programme of the school scenario in the vignette.
Report ...
}
\keyword{datasets}
