% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/its_plot.R
\name{its_plot}
\alias{its_plot}
\title{its_plot}
\usage{
its_plot(
  data_with_predictions,
  intervention_dates = c(31, 61),
  project_pre_intervention_trend = TRUE,
  colours,
  se = TRUE,
  point_shape = 3,
  point_size = 1,
  linetype = 1,
  caption = waiver(),
  title = waiver(),
  subtitle = waiver(),
  x_axis = waiver(),
  y_axis = waiver()
)
}
\arguments{
\item{data_with_predictions}{A data frame containing the initial time series data along with predicts created from `generate_predictions()`}

\item{intervention_dates}{A vector of time points (matching `time_var`) when interventions start. These time points are mutually exclusive and should not overlap. The argument accepts up to 9 scalar values representing the intervention start times. Should match `intervention_dates` argument used in `fit_its_model()`.}

\item{project_pre_intervention_trend}{Logical value whether to include a projection of the pre-intervention predicted values. Defaults to `TRUE`.}

\item{colours}{Colours passed to the `values` argument in `scale_color_manual()` and `scale_fill_manual()`. If no colours are given, defaults to `c("#3969B5", "#46C3AE")`.}

\item{se}{Logical value whether to include standard error values of the predictions. Defaults to `TRUE`.}

\item{point_shape}{Parameter passed to `shape` in `geom_point` to represent the shape of the treatment data points. Defaults to `3`.}

\item{point_size}{Parameter passed to `size` in `geom_point` to represent the size of the treatment data points. Defaults to `1`.}

\item{linetype}{Parameter passed to `linetype` in `geom_vline` to represent the line type of the vertical intervention break points. Defaults to `1`.}

\item{caption}{Optional argument passed to caption in `labs()`. If no argument is given, defaults to a few descriptive sentences on the lines shown in the plot.}

\item{title}{Optional argument passed to title in `labs()`.}

\item{subtitle}{Optional argument passed to subtitle in `labs()`.}

\item{x_axis}{Optional argument passed to x in `labs()`.}

\item{y_axis}{Optional argument passed to y in `labs()`.}
}
\value{
A ggplot object
}
\description{
Generates a ggplot2 with the values used in the ITS model along with predicted values.
}
\examples{
transform_data(data, time_var = 'time', group_var = 'group', intervention_dates = c(31, 61))
}
