% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleITScontrol-package.R
\docType{package}
\name{multipleITScontrol-package}
\alias{multipleITScontrol}
\alias{multipleITScontrol-package}
\title{multipleITScontrol: Interrupted Time Series with a Control and Multiple Interventions Through a Generalised Least Squares Framework}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions related to performing interrupted-time series through a generalised least squares (GLS) framework on linear outcomes. Allows for multiple interventions and a control with ARMA (autoregressive and moving-average) correction.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://herts-phei.github.io/multipleITScontrol/}
}

}
\author{
\strong{Maintainer}: Victor Yu \email{victor.yu@hertfordshire.gov.uk}

Other contributors:
\itemize{
  \item Emily Dickson [contributor]
}

}
\keyword{internal}
