% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_difference.R
\name{slope_difference}
\alias{slope_difference}
\title{slope_difference}
\usage{
slope_difference(model, intervention, return = TRUE)
}
\arguments{
\item{model}{model output object from \code{multipleITScontrol::fit_its_model}}

\item{intervention}{which intervention to test significance in difference. Integer value of 1, 2, or 3.}

\item{return}{Logical for returning pretty output in console. TRUE by default}
}
\value{
A transformed data frame to be passed to \code{transformed_df} in \link{fit_its_model}.
}
\description{
Ascertains whether there is a statistically significant difference in a slope or level during an intervention period between the pilot group and the control group. The estimated difference is relative to the control group. I.e. a positive coefficient means the slope of the pilot group is higher than the control group.

The result is functionally equivalent to a two sided t-test.
}
