% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_data.R
\name{transform_data}
\alias{transform_data}
\title{transform_data}
\usage{
transform_data(df, time_var, group_var, outcome_var, intervention_dates)
}
\arguments{
\item{df}{A data frame containing the initial time series data.}

\item{time_var}{A variable indicating the time index in the data frame. It must be a sequential time-series of equal intervals in numeric or a date/POSIXct/POSIXlt class.}

\item{group_var}{A character or factor variable indicating treatment and control groups. Only 'treatment' and 'control' are valid elements.}

\item{outcome_var}{A numeric variable indicating outcome.response variable.}

\item{intervention_dates}{A vector of time points (matching type of \code{time_var}) when interventions start. These time points are mutually exclusive and should not overlap. The argument accepts up to nine values representing the intervention start times.}
}
\value{
A transformed data frame to be passed to \code{transformed_df} in \link{fit_its_model}.
}
\description{
Transforms a data frame ready for input into \link{fit_its_model}, creating relevant columns for slope and level effects of interventions.

Requires a minimum of three time points in each intervention period and pre-intervention period.
}
\examples{
transform_data(data, time_var = 'time_xxx', group_var = 'group', outcome_var = 'outcome', intervention_dates = c(31, 61))
}
