% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipleITScontrol-package.R
\docType{package}
\name{multipleITScontrol-package}
\alias{multipleITScontrol}
\alias{multipleITScontrol-package}
\title{multipleITScontrol: Interrupted Time Series with Controls and Multiple Interventions through a Generalised Least Squares framework}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions related to performing interrupted-time series through a generalised least squares (GLS) framework. Allows for multiple interventions and controls.
}
\author{
\strong{Maintainer}: Victor Yu \email{victoryu@sent.com}

}
\keyword{internal}
