% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slope_difference.R
\name{slope_difference}
\alias{slope_difference}
\title{slope_difference}
\usage{
slope_difference(glm_its, output_format = "list")
}
\arguments{
\item{glm_its}{A `glm()` object representing the interrupted time series model.}

\item{output_format}{A character scalar indicating output format. Only 'list' and 'df' are valid inputs. Defaults to 'list'.}
}
\value{
A list or data.frame object containing the slope effect size difference and the corresponding 95% confidence intervals for the interventions in the ITS control model.
}
\description{
Calculates the effect size of the difference in slopes between the pilot and control groups during intervention periods, along with 95% confidence intervals.
}
\examples{
slope_difference(my_its_model)
}
